refactor components into folders for CSS modules
