+++ babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
+  plugins: ['babel-plugin-root-import']
+};
