+++ src/pages/components/HomeSection/styles.module.css
@@ -1,0 +1,33 @@
+.logoContainer {
+    padding: 0 20px 0;
+    text-align: center;
+  }
+  
+  .logoContainer h2 {
+    font-size: 300%;
+    line-height: 1em;
+    margin-top: 20px;
+    padding: 10px 0;
+    text-align: center;
+  }
+  
+  .logoContainer p {
+    margin: 0 auto;
+    max-width: 600px;
+    padding: 0.8em 0;
+    text-align: left;
+  }
+  
+  .logoContainer .logos {
+    align-items: center;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    padding: 0 20px 20px;
+  }
+  
+  .logoContainer .logos img {
+    height: 90px;
+    margin: 20px 20px 0;
+    width: auto;
+  }
