+++ src/pages/components/Splash/SplashBox.jsx
