+++ src/pages/components/Splash/SplashCircle.jsx
