+++ src/pages/components/Splash/SplashImage.jsx
