+++ src/pages/components/Splash/TypingTagline.jsx
