+++ src/pages/components/Stats/StatSlotMachineItem.jsx
