+++ src/pages/index.js
@@ -6,10 +6,10 @@ import "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./styles.module.css";
 
-import * as PageContent from "../pageContent";
-import Splash from "./components/Splash";
-import HomeSection from "./components/HomeSection";
-import StatContainer from "./components/StatContainer";
+import * as PageContent from "~/src/pageContent";
+import Splash from "./components/Splash/Splash";
+import HomeSection from "./components/HomeSection/HomeSection";
+import StatContainer from "./components/Stats/StatContainer";
 
 
 export default function Home() {
